requirements.txt
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.4.0
gunicorn==20.1.0
psycopg2-binary==2.9.1
